Mechanical code quality cleanup (PMD 123->4, SpotBugs 8->4)
Remove 37 unnecessary (int) casts left from long-to-int refactor,
move 38 misplaced field declarations to class tops, extract 19
assignment-in-operand patterns, simplify 7 boolean returns, split
4 multi-variable declarations, fix misc PMD/SpotBugs (unused fields,
log guards, loose coupling, missing @Override, premature declarations,
suboptimal collection sizing, magic numbers).

+++ src/main/java/com/mercuriusxeno/goo/block/GooCauldronInteractions.java
@@ -87,7 +87,7 @@ public final class GooCauldronInteractions {
      * @param volume volume in microblobs
      */
     private static void placeGooFluid(Level level, BlockPos pos, GooType type, int volume) {
-        int blobs = (int) Math.min(volume / MICROBLOBS_PER_BLOB, BLOBS_PER_BLOCK);
+        int blobs = Math.min(volume / MICROBLOBS_PER_BLOB, BLOBS_PER_BLOCK);
         if (blobs <= 0) { return; }
         int fluidLevel = BLOBS_PER_BLOCK - blobs;
         BlockState fluidState = GooBlocks.FLUID_BLOCKS.get(type).get()
